backend/add-cart/requirements.txt
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.2
cffi==1.15.1
click==8.1.6
cryptography==41.0.3
exceptiongroup==1.1.2
fastapi==0.101.0
h11==0.14.0
hiredis==2.2.3
httptools==0.6.0
idna==3.4
more-itertools==9.1.0
pycparser==2.21
pydantic==2.0.3
pydantic_core==2.3.0
python-decouple==3.8
python-dotenv==1.0.0
python-ulid==1.1.0
PyYAML==6.0.1
redis==4.6.0
redis-om==0.2.1
sniffio==1.3.0
starlette==0.27.0
types-pyOpenSSL==23.2.0.2
types-redis==4.6.0.3
typing_extensions==4.7.1
uvicorn==0.23.2
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
